Adding new version of NATS
