+++ docker/docker-compose.yml
@@ -1,19 +1,4 @@
 services:
-  trackmania:
-    image: evoesports/trackmania
-    restart: always
-    ports:
-      - 2351:2350/udp
-      - 2351:2350/tcp
-      - 5000:5000/tcp
-      - 5000:5000/udp
-    environment:
-      - TM_MASTERSERVER_LOGIN=testingdings
-      - TM_MASTERSERVER_PASSWORD=G1A@`SkMMfY0.vGh
-      - TM_SYSTEM_XMLRPC_ALLOWREMOTE=True
-    volumes:
-      - UserData:/server/UserData
-    
   postgres:
     image: postgres:latest
     restart: always
@@ -34,24 +19,24 @@ services:
       - 8222:8222
       - 6922:6922
         
-  evosc_sharp:
-    image: evoscsharp
-    build:
-      context: .
-      dockerfile: ../Dockerfile
-    restart: always
-    depends_on:
-      - trackmania
-      - postgres
-    environment:
-      - EVOSC_DATABASE_HOST=postgres
-      - EVOSC_DATABASE_NAME=evosc_sharp
-      - EVOSC_DATABASE_USERNAME=evosc_sharp
-      - EVOSC_DATABASE_PASSWORD=evosc_sharp
-      - EVOSC_SERVER_HOST=trackmania
-      - EVOSC_PATH_MAPS=/server/UserData/Maps
-    volumes:
-      - UserData:/server/UserData
+  #evosc_sharp:
+  #  image: evoscsharp
+  #  build:
+  #    context: .
+  #    dockerfile: ../Dockerfile
+  #  restart: always
+  #  depends_on:
+  #    - trackmania
+  #    - postgres
+    #environment:
+     # - EVOSC_DATABASE_HOST=postgres
+     # - EVOSC_DATABASE_NAME=evosc_sharp
+     # - EVOSC_DATABASE_USERNAME=evosc_sharp
+     # - EVOSC_DATABASE_PASSWORD=evosc_sharp
+     # - EVOSC_SERVER_HOST=trackmania
+  #    - EVOSC_PATH_MAPS=/server/UserData/Maps
+   # volumes:
+    #  - UserData:/server/UserData
 
 volumes:
   PostgresData: null
