+++ src/Modules/ServerSyncModule/Interfaces/IKeyValueStoreService.cs
@@ -1,4 +1,5 @@
 using NATS.Client.KeyValue;
+using NATS.Client.KeyValueStore;
 
 namespace EvoSC.Modules.EvoEsports.ServerSyncModule.Interfaces;
 
@@ -10,7 +11,7 @@ public interface IKeyValueStoreService
     /// <param name="key"></param>
     /// <param name="value"></param>
     /// <returns>Revision number of the key</returns>
-    ulong CreateEntry(string key, byte[] value);
+    ValueTask<ulong> CreateEntryAsync<T>(string key, T value);
 
     /// <summary>
     /// Create a new key value entry if one does not exist, if it does it updates.
@@ -18,7 +19,7 @@ public interface IKeyValueStoreService
     /// <param name="key"></param>
     /// <param name="value"></param>
     /// <returns>Revision number of the key</returns>
-    ulong CreateOrUpdateEntry(string key, byte[] value);
+    ValueTask<ulong> CreateOrUpdateEntryAsync<T>(string key, T value);
     
     /// <summary>
     /// Updates an existing key value entry.
@@ -26,7 +27,7 @@ public interface IKeyValueStoreService
     /// <param name="key"></param>
     /// <param name="value"></param>
     /// <param name="revision"></param>
-    void UpdateEntry(string key, byte[] value, ulong revision);
+    ValueTask<ulong> UpdateEntryAsync<T>(string key, T value, ulong revision);
     
     /// <summary>
     /// Gets an existing key value entry.
@@ -34,12 +35,12 @@ public interface IKeyValueStoreService
     /// <param name="key"></param>
     /// <param name="revision"></param>
     /// <returns>Key and value of the entry</returns>
-    KeyValueEntry GetEntry(string key, ulong revision);
+    ValueTask<NatsKVEntry<T>> GetEntryAsync<T>(string key, ulong revision);
     
     /// <summary>
     /// Deletes an existing key value entry.
     /// </summary>
     /// <param name="key"></param>
     /// <param name="revision"></param>
-    void DeleteEntry(string key, ulong revision);
+    ValueTask DeleteEntryAsync(string key, ulong revision);
 }
