+++ src/Modules/ServerSyncModule/ServerSyncModule.csproj
@@ -10,7 +10,6 @@
     <Description>Synchronize multiple servers.</Description>
     <Version>1.0.0</Version>
     <Authors>EvoEsports</Authors>
-    <GenerateTargetFrameworkAttribute>false</GenerateTargetFrameworkAttribute>
   </PropertyGroup>
   <ItemGroup>
         <EmbeddedResource Include="Templates\**\*" />
@@ -20,15 +19,11 @@
   </ItemGroup>
   <ItemGroup>
       <ProjectReference Include="../../EvoSC.Modules/EvoSC.Modules.csproj" Private="true" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="info.toml">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
+      <ProjectReference Include="../../EvoSC.Modules.SourceGeneration/EvoSC.Modules.SourceGeneration.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="NATS.Client" Version="1.1.7" />
+    <PackageReference Include="NATS.Net" Version="2.5.16" />
   </ItemGroup>
 </Project>
