+++ src/Modules/ServerSyncModule/Settings/INatsSettings.cs
@@ -16,6 +16,10 @@ public interface INatsSettings
 
     [Option(DefaultValue = "EvoSC"), Description("The name of the stream to use for state messages.")]
     public string StreamName { get; set; }
+    
+    [Option(DefaultValue = "EvoSC-consumer")]
+    [Description("The name of the consumer to use for message consumption.")]
+    public string ConsumerName { get; set; }
 
     [Option(DefaultValue = "ServerSync")]
     [Description("The message group this server is part of. Used to get messages from all other servers in the same group.")]
@@ -24,6 +28,10 @@ public interface INatsSettings
     [Option(DefaultValue = (ulong)0)]
     [Description("The last player state update message ID.")]
     public ulong PlayerStatesStartSequence { get; set; }
+    
+    [Option(DefaultValue = false)]
+    [Description("Whether to use a Key Value store for the NATS server.")]
+    public bool UseKeyValueStore { get; set; }
     
     [Option(DefaultValue = "EvoSC-bucket")]
     [Description("The name of the key vault bucket.")]
