+++ src/Modules/ServerSyncModule/info.toml
@@ -1,11 +1,11 @@
 [info]
-# A unique name for this module, this is used as a identifier
-name = "ServerSyncModule"
+# A unique name for this module, this is used as an identifier
+name="ServerSyncModule"
 # The title of the module
-title = "Server Synchronization"
+title="Server Synchronization"
 # A short description of what the module is and does
-summary = "Synchronize multiple servers."
+summary="Synchronize multiple servers."
 # The current version of this module, using SEMVER
-version = "1.0.0"
+version="1.0.0"
 # The name of the author that created this module
-author = "EvoEsports"
+author="EvoEsports"
